Correct Level, Coin Collision Coin Collection, and Restart Button (not working yet)
Did a bit of clean up. Deleted a few items from scene and changed scene to correct one so user can see updated game when using start screen. Added in collision detection for coins to be added to inventory. Worked on implementing restart button, couldn't get it working so UI is there but script additions are still in the works.